перенес всё кроме моделей в приложение api
