+++ api_yamdb/api/mixins.py
@@ -1,9 +1,15 @@
+from django.contrib import admin
 from rest_framework.mixins import (
     ListModelMixin, CreateModelMixin, DestroyModelMixin
 )
 from rest_framework.viewsets import GenericViewSet
 
 from .permissions import IsAdminOrReadOnly
+
+
+class CustomEmptyMixin(admin.ModelAdmin):
+    """Кастомный миксин для пустого поля."""
+    empty_value_display = '-пусто-'
 
 
 class CreateListDestroyViewSet(ListModelMixin,
