+++ api_yamdb/api/permissions.py
@@ -1,0 +1,25 @@
+from rest_framework import permissions
+from rest_framework.permissions import BasePermission, SAFE_METHODS
+
+
+class IsAdminOrStaff(permissions.BasePermission):
+    def has_permission(self, request, view):
+        return (request.user.is_authenticated
+                and (request.user.is_admin or request.user.is_staff))
+
+
+class IsAdminOrReadOnly(BasePermission):
+    message = 'Нужны права Администратора'
+
+    def has_permission(self, request, view):
+        return (request.method in permissions.SAFE_METHODS
+                or (request.user.is_authenticated and request.user.is_admin))
+
+
+class IsAuthorModeratorAdminSuperuser(BasePermission):
+    def has_object_permission(self, request, view, obj):
+        return (request.method in SAFE_METHODS
+                or request.user.is_admin
+                or request.user.is_moderator
+                or request.user.is_superuser
+                or obj.author == request.user)
