+++ api_yamdb/api/validators.py
@@ -2,7 +2,6 @@ from rest_framework import serializers
 
 
 class NotMeValidator:
-
     def __call__(self, new_user):
         if new_user.get('username') == 'me':
             raise serializers.ValidationError(
