+++ api_yamdb/api_yamdb/urls.py
@@ -4,9 +4,7 @@ from django.views.generic import TemplateView
 
 urlpatterns = [
     path('admin/', admin.site.urls),
-    path('api/', include('reviews.urls')),
-    path('api/', include('users.urls')),
-    path('api/', include('titles.urls')),
+    path('api/', include('api.urls')),
     path(
         'redoc/',
         TemplateView.as_view(template_name='redoc.html'),
