--- api_yamdb/reviews/serializers.py
@@ -1,40 +1,0 @@
-from rest_framework import serializers
-from rest_framework.generics import get_object_or_404
-
-from reviews.models import Comment, Review
-from titles.models import Title
-
-
-class ReviewsSerializer(serializers.ModelSerializer):
-    author = serializers.SlugRelatedField(
-        read_only=True,
-        slug_field='username'
-    )
-
-    class Meta:
-        fields = ('id', 'text', 'author', 'score', 'pub_date')
-        model = Review
-
-    def validate(self, data):
-        request = self.context['request']
-        if request.method != 'POST':
-            return data
-        current_title = get_object_or_404(
-            Title, pk=request.parser_context['kwargs'].get('title_id')
-        )
-        if current_title.reviews.filter(author=request.user).exists():
-            raise serializers.ValidationError(
-                'Вы не можете добавить более одного отзыва на произведение'
-                )
-        return data
-
-
-class CommentsSerializer(serializers.ModelSerializer):
-    author = serializers.SlugRelatedField(
-        read_only=True,
-        slug_field='username'
-    )
-
-    class Meta:
-        fields = ('id', 'text', 'author', 'pub_date')
-        model = Comment
