+++ api_yamdb/titles/models.py
@@ -1,12 +1,13 @@
 from django.db import models
+from django.conf import settings
 
 from .validators import validate_year
 
 
 class BaseModel(models.Model):
     """Базовый класс для категорий и жанров."""
-    name = models.CharField(max_length=256)
-    slug = models.SlugField(unique=True, max_length=50)
+    name = models.CharField(max_length=settings.MAX_LENGTH_TITLE_NAME)
+    slug = models.SlugField(unique=True, max_length=settings.MAX_LENGTH_SLUG)
 
     def __str__(self):
         return self.name
@@ -30,9 +31,11 @@ class Genres(BaseModel):
 
 class Title(models.Model):
     """Произведения, к которым пишут отзывы."""
-    name = models.CharField(max_length=256)
+    name = models.CharField(max_length=settings.MAX_LENGTH_TITLE_NAME)
     year = models.IntegerField(validators=[validate_year])
-    description = models.CharField(max_length=256, null=True)
+    description = models.CharField(
+        max_length=settings.MAX_LENGTH_DESCRIPTION, null=True
+    )
     genre = models.ManyToManyField(
         Genres, related_name="titles",
         blank=True
