--- api_yamdb/titles/permissions.py
@@ -1,10 +1,0 @@
-from rest_framework import permissions
-from rest_framework.permissions import BasePermission
-
-
-class IsAdminOrReadOnly(BasePermission):
-    message = 'Нужны права Администратора'
-
-    def has_permission(self, request, view):
-        return (request.method in permissions.SAFE_METHODS
-                or (request.user.is_authenticated and request.user.is_admin))
