--- api_yamdb/titles/serializers.py
@@ -1,49 +1,0 @@
-from rest_framework import serializers
-from rest_framework.relations import SlugRelatedField
-
-from .models import Categories, Genres, Title
-from .validators import validate_year
-
-
-class CategoriesSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Categories
-        exclude = ("id",)
-        lookup_field = 'slug'
-
-
-class GenresSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Genres
-        exclude = ("id",)
-        lookup_field = 'slug'
-
-
-class TitlesSerializer(serializers.ModelSerializer):
-    genre = GenresSerializer(many=True, )
-    category = CategoriesSerializer()
-    rating = serializers.IntegerField()
-
-    class Meta:
-        model = Title
-        read_only_fields = ('id', 'name', 'year', 'rating',
-                            'description', 'genre', 'category')
-        fields = ('id', 'name', 'year', 'rating',
-                  'description', 'genre', 'category')
-
-
-class TitlesCreateSerializer(serializers.ModelSerializer):
-    genre = SlugRelatedField(
-        slug_field='slug',
-        queryset=Genres.objects.all(),
-        many=True,
-    )
-    category = SlugRelatedField(
-        slug_field='slug',
-        queryset=Categories.objects.all(),
-    )
-    year = serializers.IntegerField(validators=[validate_year])
-
-    class Meta:
-        model = Title
-        fields = ('id', 'name', 'year', 'description', 'genre', 'category')
