--- api_yamdb/users/permissions.py
@@ -1,7 +1,0 @@
-from rest_framework import permissions
-
-
-class IsAdminOrStaff(permissions.BasePermission):
-    def has_permission(self, request, view):
-        return (request.user.is_authenticated
-                and (request.user.is_admin or request.user.is_staff))
